backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer-support-rag-backend"
version = "0.1.0"
description = "FastAPI backend for a local RAG customer support chatbot"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.29,<1.0",
  "langchain==0.1.20",
  "langchain-community==0.0.38",
  "sentence-transformers==2.6.1",
  "faiss-cpu==1.13.2",
  "pypdf==6.9.2",
  "python-dotenv>=1.0,<2.0",
]

[tool.setuptools]
packages = ["backend", "backend.app", "backend.scripts"]

